vk: rt: move all kusochki pieces into a single module
Stream: E400, 01:00:00 - 01:33:33
